restore README.md, more module refactor
